+++ app/src/main/java/com/example/talkify/utils/Category.kt
@@ -1,0 +1,3 @@
+package com.example.talkify.utils
+
+data class Category(val name: String, val Imageid: Int)
