+++ app/src/main/java/com/example/talkify/utils/CategoryList.kt
@@ -1,0 +1,12 @@
+package com.example.talkify.utils
+
+import com.example.talkify.R
+
+
+val categories = listOf<Category>(
+    Category("Fruits", R.drawable.img_6),
+
+    Category("Cars", R.drawable.cars),
+    Category("Emotions", R.drawable.emotions),
+    Category("Custom", R.drawable.custom)
+)
