+++ app/src/main/java/com/example/talkify/utils/Item.kt
@@ -1,0 +1,3 @@
+package com.example.talkify.utils
+
+data class Item(val name: String, val Imageid: Int)
