+++ app/src/main/java/com/example/talkify/utils/ItemsList.kt
@@ -1,0 +1,41 @@
+package com.example.talkify.utils
+
+import com.example.talkify.R
+
+
+val items = listOf<Item>(
+
+    Item("1", R.drawable.),
+    Item("1", R.drawable.img_2),
+    Item("1", R.drawable.img_3),
+    Item("1", R.drawable.img_4),
+    Item("1", R.drawable.img_5),
+    Item("1", R.drawable.img_6),
+    Item("1", R.drawable.img_2),
+    Item("1", R.drawable.img_3),
+    Item("1", R.drawable.img_4),
+    Item("1", R.drawable.img_5),
+    Item("1", R.drawable.img_6),
+    Item("1", R.drawable.img),
+    Item("1", R.drawable.img_2),
+    Item("1", R.drawable.img_3),
+    Item("1", R.drawable.img_4),
+    Item("1", R.drawable.img_5),
+    Item("1", R.drawable.img_6),
+    Item("1", R.drawable.img_2),
+    Item("1", R.drawable.img_3),
+    Item("1", R.drawable.img_4),
+    Item("1", R.drawable.img_5),
+    Item("1", R.drawable.img_6),
+    Item("1", R.drawable.img),
+    Item("1", R.drawable.img_2),
+    Item("1", R.drawable.img_3),
+    Item("1", R.drawable.img_4),
+    Item("1", R.drawable.img_5),
+    Item("1", R.drawable.img_6),
+    Item("1", R.drawable.img_2),
+    Item("1", R.drawable.img_3),
+    Item("1", R.drawable.img_4),
+    Item("1", R.drawable.img_5),
+    Item("1", R.drawable.img_6),
+)
